pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include the hydra_plugins namespace package in the wheel.
# No src/ layout — Hydra namespace discovery requires the package
# to be directly importable after installation.
packages = ["hydra_plugins"]

[project]
name = "hydra-apptainer-launcher"
version = "0.1.0"
description = "Hydra submitit launcher with Apptainer support for HPC clusters"
requires-python = ">=3.10"

dependencies = [
    # Provides BaseSubmititLauncher, SlurmQueueConf, BaseQueueConf.
    # submitit>=1.3.3 is pulled in transitively.
    "hydra-submitit-launcher>=1.2.0",
]
